requirements.txt
fastapi
uvicorn
pinecone-client
openai
pandas
python-dotenv
pydantic